+++ examples/with-nextjs/frontend/package.json
@@ -8,9 +8,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@ponder/client": "workspace:*",
     "@tanstack/react-query": "^5.12.2",
-    "graphql": "^16.8.1",
-    "graphql-request": "^6.1.0",
     "next": "14.0.3",
     "react": "^18",
     "react-countup": "^6.5.0",
