+++ examples/with-nextjs/frontend/src/hooks/index.ts
@@ -1,0 +1,6 @@
+import { createClient } from "@ponder/client";
+import * as schema from "../../../ponder/ponder.schema";
+
+const client = createClient("http://localhost:42069", { schema });
+
+export { client, schema };
