+++ examples/with-nextjs/ponder/src/api/index.ts
@@ -1,0 +1,8 @@
+import { Hono } from "hono";
+import { client } from "ponder";
+
+const app = new Hono();
+
+app.use(client());
+
+export default app;
