+++ packages/core/src/index.ts
@@ -30,7 +30,7 @@ export {
   primaryKey,
   hex,
   bigint,
-} from "@/drizzle/index.js";
+} from "@/drizzle/onchain.js";
 
 export { client } from "@/client/index.js";
