+++ packages/core/src/indexing-store/historical.ts
@@ -13,9 +13,9 @@ import {
   type Schema,
   getPrimaryKeyColumns,
   getTableNames,
-  onchain,
 } from "@/drizzle/index.js";
 import { getColumnCasing } from "@/drizzle/kit/index.js";
+import { onchain } from "@/drizzle/onchain.js";
 import { encodeCheckpoint, zeroCheckpoint } from "@/utils/checkpoint.js";
 import { prettyPrint } from "@/utils/print.js";
 import { createQueue } from "@ponder/common";
