+++ packages/core/src/types/virtual.test-d.ts
@@ -1,5 +1,5 @@
 import { createConfig } from "@/config/config.js";
-import { onchainTable } from "@/drizzle/index.js";
+import { onchainTable } from "@/drizzle/onchain.js";
 import { http, type Abi, type Address, type Hex, parseAbiItem } from "viem";
 import { assertType, test } from "vitest";
 import type { Db } from "./db.js";
