+++ packages/core/tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   name: "ponder",
-  entry: ["src/index.ts", "src/bin/ponder.ts"],
+  entry: ["src/index.ts", "src/bin/ponder.ts", "src/drizzle/onchain.ts"],
   outDir: "dist",
   format: ["esm"],
   sourcemap: true,
